requirements.txt
Django==4.2.10
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
psycopg2==2.9.9
drf-yasg==1.21.7
django-cors-headers==4.3.1